requirements.txt
aiogram==3.0.0b7
openai-whisper
ffmpeg-python
